x/escrow/types/message_buyer_dispute_rebuttal_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/crow-labs/delta/testutil/sample"
	"github.com/stretchr/testify/require"
)

func TestMsgBuyerDisputeRebuttal_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgBuyerDisputeRebuttal
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgBuyerDisputeRebuttal{
				Buyer: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgBuyerDisputeRebuttal{
				Buyer: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
